DraftBot: fold partnership % into stats_core's team-draft formula
Simplify pass: one formula, one owner -- the Vault/Key wrapper duplicated
calculate_team_draft_win_percentage under a third name. Also hoist the
test_db fixture into tests/conftest.py for new test files (older files
keep their shadowing local copies).
